LayoutTests/fast/events/no-mousemove-on-modifier-key-press-expected.txt

This test verifies that MouseMove events are not fired when modifier keys are pressed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


To run this test by hand, press a modifier key.

Dispatched KeyDown event with modifier metaKey
Dispatched KeyDown event with modifier altKey
Dispatched KeyDown event with modifier shiftKey
Dispatched KeyDown event with modifier ctrlKey
PASS successfullyParsed is true

TEST COMPLETE
